requirements.txt
joblib==1.2.0
numpy==1.23.1
pytest==7.2.1
python-dotenv==1.0.0
scientific_experiment_manager==0.1.0
scipy==1.8.1
sem==0.3.8
